Re #1700 Renamed project for testing fileParameters to file_parameters instead of file_parameters_tester
Mainly to kick on the build
