+++ _LowLevelCode/cpp/combine_sqw/sqw_reader.h
@@ -2,7 +2,7 @@
 #define H_SQW_READER
 
 #include "pix_mem_map.h"
-#include "../file_parameters_tester/fileParameters.h"
+#include "../file_parameters/fileParameters.h"
 //-----------------------------------------------------------------------------------------------------------------
 class sqw_reader
 {
